[CORE-1121] Correct pachctl delete repo --all for projects (#8393)
* `pachctl delete repo --all` now calls a new gRPC RPC `DeleteRepos` which attempts to delete all repos in the current project. It is not an error if one or more repos may not be deleted; they simply will not be.

+++ src/server/pfs/server/api_server.go
@@ -11,6 +11,8 @@ import (
 	"github.com/gogo/protobuf/jsonpb"
 	"github.com/gogo/protobuf/proto"
 	"github.com/gogo/protobuf/types"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 
 	"github.com/pachyderm/pachyderm/v2/src/auth"
 	"github.com/pachyderm/pachyderm/v2/src/client"
@@ -162,12 +164,29 @@ func (a *apiServer) DeleteRepoInTransaction(txnCtx *txncontext.TransactionContex
 // DeleteRepo implements the protobuf pfs.DeleteRepo RPC
 func (a *apiServer) DeleteRepo(ctx context.Context, request *pfs.DeleteRepoRequest) (response *types.Empty, retErr error) {
 	request.GetRepo().EnsureProject()
+	if request.GetRepo() == nil {
+		return nil, status.Error(codes.InvalidArgument, "no repo specified")
+	}
 	if err := a.env.TxnEnv.WithTransaction(ctx, func(txn txnenv.Transaction) error {
 		return errors.EnsureStack(txn.DeleteRepo(request))
 	}, nil); err != nil {
 		return nil, err
 	}
 	return &types.Empty{}, nil
+}
+
+// DeleteRepos implements the pfs.DeleteRepo RPC.  It deletes more than one repo at once.
+func (a *apiServer) DeleteRepos(ctx context.Context, request *pfs.DeleteReposRequest) (resp *pfs.DeleteReposResponse, err error) {
+	if len(request.Projects) == 0 {
+		return nil, status.Error(codes.InvalidArgument, "DeleteRepos must specify project(s) whose repos should be deleted.")
+	}
+	repos, err := a.driver.deleteProjectsRepos(ctx, request.Projects)
+	if err != nil {
+		return nil, err
+	}
+	return &pfs.DeleteReposResponse{
+		Repos: repos,
+	}, nil
 }
 
 // StartCommitInTransaction is identical to StartCommit except that it can run
